+++ backend/internal/infrastructure/http/router/router.go
@@ -1,4 +1,4 @@
-package http
+package router
 
 import (
 	"myapp/internal/controller/api"
