+++ backend/main.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"myapp/internal/config"
 	"myapp/internal/infrastructure/database"
-	"myapp/internal/infrastructure/http"
-	"myapp/internal/middleware"
+	"myapp/internal/infrastructure/http/middleware"
+	"myapp/internal/infrastructure/http/router"
 
 	"github.com/gin-gonic/gin"
 )
@@ -18,6 +18,6 @@ func main() {
 	app := gin.Default()
 	app.Use(middleware.Transaction())
 	app.Use(middleware.Cors())
-	http.SetupRoutes(app)
+	router.SetupRoutes(app)
 	app.Run(fmt.Sprintf("%s:%d", config.HostName, config.Port))
 }
